requirements.txt
fastapi
uvicorn
psycopg2
SQLAlchemy
python-dotenv
alembic
passlib[bcrypt]
python-jose[cryptography]
python-multipart
argon2
